Address live preview review comments
Move the live-preview panel bottom padding inside the conditional panel so liveTranscriptPreview is not recomputed during body layout.

Replace the duplicate dictation preview drain gate with a shared Core one-shot void continuation while keeping the scheduler's Bool continuation concrete for Swift 6 sendability.

Verification:

- swift test --filter STTSchedulerTests

- swift test --filter DictationServiceTests

- MACPARAKEET_SKIP_WHISPERKIT=1 swift build --build-path .build-swift6-no-whisper -Xswiftc -swift-version -Xswiftc 6

- git diff --check
