+++ Sources/MacParakeetCore/Services/Dictation/DictationService.swift
@@ -79,22 +79,7 @@ private struct DictationDisplayPreviewState: Sendable {
     let task: Task<Void, Never>
 }
 
-private final class DictationPreviewDrainGate: @unchecked Sendable {
-    private let lock = NSLock()
-    private var continuation: CheckedContinuation<Void, Never>?
-
-    init(_ continuation: CheckedContinuation<Void, Never>) {
-        self.continuation = continuation
-    }
-
-    func resume() {
-        lock.lock()
-        let continuation = continuation
-        self.continuation = nil
-        lock.unlock()
-        continuation?.resume()
-    }
-}
+private typealias DictationPreviewDrainGate = OneShotVoidContinuation
 
 public enum AIFormatterAppContextPhase: Sendable {
     case start
